[UE] make text block UE editable
- remove aem.js and update lib-franklin.js from latest aem.js
