+++ scripts/editor-support.js
@@ -6,7 +6,7 @@ import {
   decorateSections,
   loadBlock,
   loadBlocks,
-} from './aem.js';
+} from './lib-franklin.js';
 import { decorateRichtext } from './editor-support-rte.js';
 import { decorateMain } from './scripts.js';
